function project
This is the project I created with the common functions. What you can
do is copy this project straight into eclipse so we have the exact same
project. The only function the project has right now is the dimensional
list function, which works quite well. I will show you how to use it
later. I also made just the .py file in my projects folder, so you have
the entire project and just the file now